NumberSprite Update
the Number Sprite graphics were unweildly, being twice the width and
height that they needed to be to fit the number.  This update accounts
for changing the size of the gfx that contains the sprite info on what
NumberSprites look like, and makes adjustments to line the title up
accordingly.